+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   webnode in Docker. ([#1778](https://github.com/o1-labs/mina-rust/pull/1778))
 - **CI**: add beta channel lint workflow to catch build issues early
   ([#1875](https://github.com/o1-labs/mina-rust/pull/1875))
+- **Feature**: Add logging if initial peers are invalid
+  ([#1703](https://github.com/o1-labs/mina-rust/pull/1703))
 
 ### Fixed